Fixed overflow bug in computation of r2Max
git-svn-id: http://svn.code.sf.net/p/petavision/code/trunk@4092 7c0444ca-5550-0410-9008-faecdbe78eb5

+++ src/connections/InitCocircWeights.cpp
@@ -116,7 +116,7 @@ bool InitCocircWeights::calcDistChordCocircKurvePreNKurvePost(
    const float shift = weightParamPtr->getshift();
    const float sigma = weightParamPtr->getsigma();
    const float sigma2 = 2 * sigma * sigma;
-   const float r2Max = weightParamPtr->getr2Max();
+   const double r2Max = weightParamPtr->getr2Max();
    const int numFlanks = weightParamPtr->getnumFlanks();
    float thetaPre = weightParamPtr->getthPre();
